+++ src/openga/applications/singleMachineOASPSD_TPSGA.java
@@ -8,53 +8,24 @@ import openga.operator.localSearch.*;
 import openga.ObjectiveFunctions.*;
 import openga.MainProgram.*;
 import openga.Fitness.*;
-import openga.applications.data.OASInstancesWithTOU;
+import openga.applications.data.OASInstances;
 
 /*
  * <p>Title: The OpenGA project which is to build general framework of Genetic algorithm.</p>
  * <p>Description: </p>
  * <p>Copyright: Copyright (c) 2016</p>
  * <p>Company: Cheng Shiu University</p>
- * @authors Chen, Shih-Hsin ; Guo, Yu-Cheng
+ * @authors Chen, Shih-Hsin ; Chang, Yu-Tang
  * @version 1.0
  * We obtain the instance from OASLIB.
  * Reference "A tabu search algorithm for order acceptance and scheduling", http://home.ku.edu.tr/~coguz/Research/Dataset_OAS.zip
- * We use an electricity data from a paper.
  */
-public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
-
-  public int numberOfSalesmen;
-  public int type; //type = 0 : mtsp,  type = 2 : OA,  type = 3 : TCX
-  int maxNeighborhood;  //A default value of the maximum neighbors to search.
-  int TournamentSize;
-  double alpha;
-
-  singleThreadGA GaMain;
-  SelectI Selection;
-  CrossoverMTSPI Crossover;
-  MutationMTSPI Mutation;
-  localSearchMTSPI localSearch1;
-  populationI Population;
-  ObjectiveFunctionOASWithTOUI[] ObjectiveFunction;
-  String instanceName = "";
-
-  boolean applyLocalSearch;
-
-  double[] r;       //  release date.
-  double[] p;       //  processing time
-  double[] d;       //  due-date
-  double[] d_bar;   //  deadline
-  double[] e;       //  revenue
-  double[] w;       //  weight
-  double[] power;   //  power
-  double[][] s;     //  setup times
-
-  public singleMachineOAS_SGAWithTOU() {
-  }
-
+public class singleMachineOASPSD_TPSGA extends singleMachineOAS_TPSGA {
+  double b; //Setup index for PSD.
+  
   public void setParameter(double crossoverRate, double mutationRate, int counter, double elitism,
           int generation, int type, int numberOfSalesmen, int cities, String instanceName,
-          double[] r, double[] p, double[] d, double[] d_bar, double[] e, double[] w, double[] power, double[][] s, int pop_Size) {
+          double[] r, double[] p, double[] d, double[] d_bar, double[] e, double[] w, double b, int pop_Size) {
     this.DEFAULT_crossoverRate = crossoverRate;
     this.DEFAULT_mutationRate = mutationRate;
     this.counter = counter;
@@ -70,17 +41,10 @@ public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
     this.d_bar = d_bar;
     this.e = e;
     this.w = w;
-    this.power = power;
-    this.s = s;
+    this.b = b;
     this.DEFAULT_PopSize = pop_Size;
-  }
+  }  
 
-  public void setLocalSearchData(boolean applyLocalSearch, int maxNeighborhood) {
-    this.applyLocalSearch = applyLocalSearch;
-    this.maxNeighborhood = maxNeighborhood;
-  }
-
-  @Override
   public void initiateVars() {
     GaMain = new singleThreadGAwithInitialPop();//singleThreadGAwithMultipleCrossover singleThreadGA adaptiveGA
     Population = new population();
@@ -88,8 +52,8 @@ public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
     Crossover = new TCSCFCrossover(type);
     Mutation = new swapMutationTwoPart();//TwoPartMTSPMutation
     localSearch1 = new localSearchByIG();
-    ObjectiveFunction = new ObjectiveFunctionOASWithTOUI[numberOfObjs];
-    ObjectiveFunction[0] = new TPObjectiveFunctionforOASWithTOU();
+    ObjectiveFunction = new ObjectiveFunctionOASI[numberOfObjs];
+    ObjectiveFunction[0] = new TPObjectiveFunctionforOAS_Setup();
     Fitness = new singleObjectiveFitness();//singleObjectiveFitness singleObjectiveFitnessByNormalize
     objectiveMinimization = new boolean[numberOfObjs];
     objectiveMinimization[0] = false;
@@ -106,11 +70,11 @@ public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
       Population.getSingleChromosome(i).generateTwoPartPop(length + numberOfSalesmen, numberOfSalesmen);
     }
 //    Population.
-    Selection.setTournamentSize(2);//Binary tournament
+    Selection.setTournamentSize(2);
     Crossover.setNumberofSalesmen(numberOfSalesmen);
     Mutation.setNumberofSalesmen(numberOfSalesmen);
     localSearch1.setNumberofSalesmen(numberOfSalesmen);
-    ObjectiveFunction[0].setOASData(r, p, d, d_bar, e, w, power, s, numberOfSalesmen);
+    ObjectiveFunction[0].setOASData(r, p, d, d_bar, e, w, b, numberOfSalesmen);
     //set the data to the GA main program.
     /*Note: the gene length is problem size + numberOfSalesmen*/
     GaMain.setData(Population, Selection, Crossover, Mutation, ObjectiveFunction, Fitness, DEFAULT_generations,
@@ -119,44 +83,42 @@ public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
     GaMain.setLocalSearchOperator(localSearch1, applyLocalSearch, maxNeighborhood);
   }
 
-  @Override
   public void start() {
     openga.util.timeClock timeClock1 = new openga.util.timeClock();
     timeClock1.start();
     GaMain.startGA();
     timeClock1.end();
-
     if (type == 3) {//3: TCX (Original) The TCX
-      type = 1;
+      type = 1;//The best in terms of the Obj and CPU time
     } else if (type == 2) {//2: Last salesmen reserve the same sites
-      type = 2;
+      type = 2;//Actually no difference compared with the TCX
     } else if (type == 0) {//0: All salesmen reserve the same sites
-      type = 3;
+      type = 3;//The worst.
     }
-    String implementResult = instanceName + "\t" + DEFAULT_crossoverRate + "\t" + DEFAULT_mutationRate + "\t" + type + "\t" + DEFAULT_PopSize + "\t" + applyLocalSearch + "\t" + alpha
+    String implementResult = instanceName + "\t" + DEFAULT_crossoverRate + "\t" + DEFAULT_mutationRate + "\t" + type + "\t" + DEFAULT_PopSize + "\t" + alpha + "\t" + b
             + "\t" + GaMain.getArchieve().getSingleChromosome(0).getObjValue()[0]
             + "\t" + timeClock1.getExecutionTime() / 1000.0 + "\n";
-//    writeFile("OASforSMSP_TOU_20171130" + "MaxRevenueFull", implementResult);
+    writeFile("OASforSMSPPSD_20180810_Setup" + "MaxRevenueFull", implementResult);
     System.out.print(implementResult);
   }
 
   public static void main(String[] args) {
-    System.out.println("OASforSMSP_TOU_20171130" + "MaxRevenueFull");
+    System.out.println("OASforSMSPPSD-SGALS_20180810" + "MaxRevenueFull");
     int counter = 0;
     boolean applyLocalSearch;
-    double[] crossoverRate = new double[]{1, 0.5};//1, 0.5 [0.5]
-    double[] mutationRate = new double[]{0.5};//0.1, 0.5 [0.5]
+    double[] crossoverRate = new double[]{1,0.5};//0.9, 0.5 [0.5]
+    double[] mutationRate = new double[]{0.1,0.5};//0.1, 0.5 [0.5]
     double elitism[] = new double[]{0.1};
-    //0: All salesmen reserve the same sites,2: Last salesmen reserve the same sites,3: TCX (Original)
-    int[] crossoverType = new int[]{0};//0, 2, 3 [0] No Obj differences; 3 is a little bit higher. 0 is fatest.
+    int[] crossoverType = new int[]{3,2};//0: All salesmen reserve the same sites,2: Last salesmen reserve the same sites,3: TCX (Original)
     int repeat = 3;//30
     int generations[] = new int[]{0};//1000
-    int[] populationsSize = new int[]{100, 200};
+    int[] populationsSize = new int[]{50, 100, 200};//50, 100, 200 [50]
     int[] numberOfSalesmen = new int[]{2};
-    double[] alpha = new double[]{0.1};//0.2, 0.1, 0.05 [0.1] //LocalSearch maxNeighborhood
+    double[] alpha = new double[]{0.2, 0.1, 0.05, 0};//0.2, 0.1, 0.05 [0.2]
     int[] orders = new int[]{10, 15, 20, 25, 50, 100};//10, 15, 20, 25, 50, 100
     int[] Tao = new int[]{1, 3, 5, 7, 9};//1, 3, 5, 7, 9
     int[] R = new int[]{1, 3, 5, 7, 9};//1, 3, 5, 7, 9
+    double b[] = new double[]{0.1, 0.2, 0.3};//Setup index for PSD.
     int instanceReplications = 1;
 
     for (int i = 0; i < orders.length; i++) {
@@ -166,9 +128,9 @@ public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
             continue;
           }
           for (int l = 0; l < instanceReplications; l++) {
-            OASInstancesWithTOU OASInstances1 = new OASInstancesWithTOU();
-            String instanceName = new String(".\\instances\\SingleMachineOASWithTOU\\" + orders[i] + "orders\\Tao" + Tao[j] + "\\R" + R[k]
-                    + "\\Dataslack_" + orders[i] + "orders_Tao" + Tao[j] + "R" + R[k] + "_" + (l + 1) + ".txt");
+            OASInstances OASInstances1 = new OASInstances();
+            String instanceName = new String("./instances/SingleMachineOAS/" + orders[i] + "orders/Tao" + Tao[j] + "/R" + R[k]
+                    + "/Dataslack_" + orders[i] + "orders_Tao" + Tao[j] + "R" + R[k] + "_" + (l + 1) + ".txt");
             OASInstances1.setData(instanceName, orders[i]);
             OASInstances1.getDataFromFile();
 
@@ -177,32 +139,33 @@ public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
                 for (int n = 0; n < mutationRate.length; n++) {
                   for (int o = 0; o < elitism.length; o++) {
                     for (int p = 0; p < numberOfSalesmen.length; p++) {
-                      for (int ls = 0; ls < 2; ls++) {
-                        if (ls == 0) {
-                          applyLocalSearch = false;
-                        } else {
-                          applyLocalSearch = true;
-                        }
                         for (int q = 0; q < alpha.length; q++) {
                           for (int pop_Size = 0; pop_Size < populationsSize.length; pop_Size++) {
-                            for (int r = 0; r < repeat; r++) {
-                              int _alpha = (int) Math.round(((double) orders[i] * alpha[q]));
-                              generations[0] = orders[i] * (numberOfSalesmen[p] - 1) * 2000 / populationsSize[pop_Size];
-                              singleMachineOAS_SGAWithTOU TSP1 = new singleMachineOAS_SGAWithTOU();
-                              TSP1.alpha = alpha[q];
-                              TSP1.setParameter(crossoverRate[m], mutationRate[n], counter, elitism[o], generations[0],
-                                      crossoverType[t], numberOfSalesmen[p], OASInstances1.getSize(), instanceName,
-                                      OASInstances1.getR(), OASInstances1.getP(), OASInstances1.getD(), OASInstances1.getD_bar(), OASInstances1.getE(), OASInstances1.getW(),
-                                      OASInstances1.getPower(), OASInstances1.getS(), populationsSize[pop_Size]);
-                              TSP1.setLocalSearchData(applyLocalSearch, _alpha);
-                              TSP1.initiateVars();
-                              TSP1.start();
-//                        TSP1.printResults();
-                              counter++;
+                            for(int s = 0 ; s < b.length ; s++){
+                              for (int r = 0; r < repeat; r++) {
+                                if (alpha[q] != 0) {
+                                  applyLocalSearch = true;
+                                } else {
+                                  applyLocalSearch = false;
+                                }                                  
+
+                                int _alpha = (int) Math.round(((double) orders[i] * alpha[q]));
+                                generations[0] = orders[i] * (numberOfSalesmen[p] - 1) * 2000 / populationsSize[pop_Size];
+                                singleMachineOASPSD_TPSGA TSP1 = new singleMachineOASPSD_TPSGA();
+                                TSP1.alpha = alpha[q];
+                                TSP1.setParameter(crossoverRate[m], mutationRate[n], counter, elitism[o], generations[0],
+                                        crossoverType[t], numberOfSalesmen[p], OASInstances1.getSize(), instanceName,
+                                        OASInstances1.getR(), OASInstances1.getP(), OASInstances1.getD(), OASInstances1.getD_bar(),
+                                        OASInstances1.getE(), OASInstances1.getW(), b[s], populationsSize[pop_Size]);
+                                TSP1.setLocalSearchData(applyLocalSearch, _alpha);
+                                TSP1.initiateVars();
+                                TSP1.start();
+  //                        TSP1.printResults();
+                                counter++;
+                              }                              
                             }
                           }
                         }
-                      }
                     }
                   }
                 }
@@ -214,28 +177,5 @@ public class singleMachineOAS_SGAWithTOU extends mTSPSGATwoPart {
       }
     }
     System.exit(0);
-  }
-
-  public void printResults() {
-    //to output the implementation result.
-    String implementResult = "";
-    implementResult = "";
-    int Clength = length + numberOfSalesmen;
-    int cities = length;
-
-    for (int k = 0; k < GaMain.getArchieve().getPopulationSize(); k++) {
-      for (int j = 0; j < numberOfObjs; j++) {//for each objectives
-        implementResult += GaMain.getArchieve().getObjectiveValues(k)[j] + "\t";
-      }
-      for (int j = 0; j < Clength; j++) {//for each objectives
-        if (j < cities) {
-          implementResult += (GaMain.getArchieve().getSingleChromosome(k).genes[j] + 1) + " ";
-        } else {
-          implementResult += (GaMain.getArchieve().getSingleChromosome(k).genes[j]) + " ";
-        }
-      }
-      implementResult += "\n";
-    }
-    writeFile("singleMachineArchive_" + Clength, implementResult);
   }
 }
