+++ src/openga/applications/singleMachineOAS_TPSGA_Co2.java
@@ -8,19 +8,20 @@ import openga.operator.localSearch.*;
 import openga.ObjectiveFunctions.*;
 import openga.MainProgram.*;
 import openga.Fitness.*;
-import openga.applications.data.OASInstances;
+import openga.applications.data.OASInstancesWithTOU;
 
 /*
  * <p>Title: The OpenGA project which is to build general framework of Genetic algorithm.</p>
  * <p>Description: </p>
  * <p>Copyright: Copyright (c) 2016</p>
  * <p>Company: Cheng Shiu University</p>
- * @authors Chen, Shih-Hsin ; Chang, Yu-Tang
+ * @authors Chen, Shih-Hsin ; Guo, Yu-Cheng
  * @version 1.0
  * We obtain the instance from OASLIB.
  * Reference "A tabu search algorithm for order acceptance and scheduling", http://home.ku.edu.tr/~coguz/Research/Dataset_OAS.zip
+ * We use an electricity data from a paper.
  */
-public class singleMachineOAS_SGA extends mTSPSGATwoPart {
+public class singleMachineOAS_TPSGA_Co2 extends mTSPSGATwoPart {
 
   public int numberOfSalesmen;
   public int type; //type = 0 : mtsp,  type = 2 : OA,  type = 3 : TCX
@@ -34,7 +35,7 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
   MutationMTSPI Mutation;
   localSearchMTSPI localSearch1;
   populationI Population;
-  ObjectiveFunctionOASI[] ObjectiveFunction;
+  ObjectiveFunctionOASWithTOUI[] ObjectiveFunction;
   String instanceName = "";
 
   boolean applyLocalSearch;
@@ -45,14 +46,15 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
   double[] d_bar;   //  deadline
   double[] e;       //  revenue
   double[] w;       //  weight
+  double[] power;   //  power
   double[][] s;     //  setup times
 
-  public singleMachineOAS_SGA() {
+  public singleMachineOAS_TPSGA_Co2() {
   }
 
   public void setParameter(double crossoverRate, double mutationRate, int counter, double elitism,
-          int generation, int type, int numberOfSalesmen, int cities, String instanceName,
-          double[] r, double[] p, double[] d, double[] d_bar, double[] e, double[] w, double[][] s, int pop_Size) {
+          int generation, int type, int numberOfSalesmen, int cities, String instanceName, int pop_Size,
+          double[] p, double[] power) {
     this.DEFAULT_crossoverRate = crossoverRate;
     this.DEFAULT_mutationRate = mutationRate;
     this.counter = counter;
@@ -62,14 +64,9 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
     this.numberOfSalesmen = numberOfSalesmen;
     this.length = cities;
     this.instanceName = instanceName;
-    this.r = r;
-    this.p = p;
-    this.d = d;
-    this.d_bar = d_bar;
-    this.e = e;
-    this.w = w;
-    this.s = s;
     this.DEFAULT_PopSize = pop_Size;
+    this.p = p;//***update***
+    this.power = power;//***update***
   }
 
   public void setLocalSearchData(boolean applyLocalSearch, int maxNeighborhood) {
@@ -85,11 +82,11 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
     Crossover = new TCSCFCrossover(type);
     Mutation = new swapMutationTwoPart();//TwoPartMTSPMutation
     localSearch1 = new localSearchByIG();
-    ObjectiveFunction = new ObjectiveFunctionOASI[numberOfObjs];
-    ObjectiveFunction[0] = new TPObjectiveFunctionforOAS();
+    ObjectiveFunction = new ObjectiveFunctionOASWithTOUI[numberOfObjs];
+    ObjectiveFunction[0] = new ObjFunctionCo2();//***update***
     Fitness = new singleObjectiveFitness();//singleObjectiveFitness singleObjectiveFitnessByNormalize
     objectiveMinimization = new boolean[numberOfObjs];
-    objectiveMinimization[0] = false;
+    objectiveMinimization[0] = true;
     encodeType = true;
 
     if (numberOfSalesmen >= length) {
@@ -103,11 +100,11 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
       Population.getSingleChromosome(i).generateTwoPartPop(length + numberOfSalesmen, numberOfSalesmen);
     }
 //    Population.
-    Selection.setTournamentSize(7);
+    Selection.setTournamentSize(2);//Binary tournament
     Crossover.setNumberofSalesmen(numberOfSalesmen);
     Mutation.setNumberofSalesmen(numberOfSalesmen);
     localSearch1.setNumberofSalesmen(numberOfSalesmen);
-    ObjectiveFunction[0].setOASData(r, p, d, d_bar, e, w, s, numberOfSalesmen);
+    ObjectiveFunction[0].setPowerData(p, power);//***update***
     //set the data to the GA main program.
     /*Note: the gene length is problem size + numberOfSalesmen*/
     GaMain.setData(Population, Selection, Crossover, Mutation, ObjectiveFunction, Fitness, DEFAULT_generations,
@@ -122,37 +119,38 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
     timeClock1.start();
     GaMain.startGA();
     timeClock1.end();
-    if (type == 3) {
+
+    if (type == 3) {//3: TCX (Original) The TCX
       type = 1;
-    } else if (type == 2) {
-      type = 2;//Significant
-    } else if (type == 0) {
-      type = 3; 
+    } else if (type == 2) {//2: Last salesmen reserve the same sites
+      type = 2;
+    } else if (type == 0) {//0: All salesmen reserve the same sites
+      type = 3;
     }
     String implementResult = instanceName + "\t" + DEFAULT_crossoverRate + "\t" + DEFAULT_mutationRate + "\t" + type + "\t" + DEFAULT_PopSize + "\t" + applyLocalSearch + "\t" + alpha
             + "\t" + GaMain.getArchieve().getSingleChromosome(0).getObjValue()[0]
             + "\t" + timeClock1.getExecutionTime() / 1000.0 + "\n";
-    writeFile("OASforSMSP_20171206-SGA" + "MaxRevenueFull", implementResult);
+    writeFile("OASforSMSP_Co2_20171130" + "MinCo2Cost", implementResult);
     System.out.print(implementResult);
   }
 
   public static void main(String[] args) {
-    System.out.println("OASforSMSP_20171206" + "MaxRevenueFull");
+    System.out.println("OASforSMSP_Co2_20171130" + "Co2Cost");
     int counter = 0;
     boolean applyLocalSearch;
-    double[] crossoverRate = new double[]{0.5};//0.9, 0.5 [0.5]
-    double[] mutationRate = new double[]{0.5};//0.1, 0.5 [0.5]
+    double[] crossoverRate = new double[]{0.5, 0.9};//1, 0.5 [0.5]
+    double[] mutationRate = new double[]{0.1, 0.5};//0.1, 0.5 [0.5]
     double elitism[] = new double[]{0.1};
-    int[] crossoverType = new int[]{2};//0: All salesmen reserve the same sites,2: Last salesmen reserve the same sites,3: TCX (Original) [2]
+    //0: All salesmen reserve the same sites,2: Last salesmen reserve the same sites,3: TCX (Original)
+    int[] crossoverType = new int[]{0};//0, 2, 3 [0] No Obj differences; 3 is a little bit higher. 0 is fatest.
     int repeat = 30;//30
     int generations[] = new int[]{0};//1000
-    int[] populationsSize = new int[]{50};//50, 100, 200 [50]
+    int[] populationsSize = new int[]{100, 200};
     int[] numberOfSalesmen = new int[]{2};
-    //alpha is the local search parameter. 0 means there is no local search applied.
-    double[] alpha = new double[]{0.1};//0.2, 0.1, 0.05, 0 [0.1]
-    int[] orders = new int[]{10, 15, 20, 25, 50, 100};//10, 15, 20, 25, 50, 100
-    int[] Tao = new int[]{1, 3, 5, 7, 9};//1, 3, 5, 7, 9
-    int[] R = new int[]{1, 3, 5, 7, 9};//1, 3, 5, 7, 9
+    double[] alpha = new double[]{0.1};//0.2, 0.1, 0.05 [0.1] //LocalSearch maxNeighborhood
+    int[] orders = new int[]{10,25,50,};//10, 15, 20, 25, 50, 100
+    int[] Tao = new int[]{5};//1, 3, 5, 7, 9
+    int[] R = new int[]{5};//1, 3, 5, 7, 9
     int instanceReplications = 1;
 
     for (int i = 0; i < orders.length; i++) {
@@ -162,9 +160,9 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
             continue;
           }
           for (int l = 0; l < instanceReplications; l++) {
-            OASInstances OASInstances1 = new OASInstances();
-            String instanceName = new String("./instances/SingleMachineOAS/" + orders[i] + "orders/Tao" + Tao[j] + "/R" + R[k]
-                    + "/Dataslack_" + orders[i] + "orders_Tao" + Tao[j] + "R" + R[k] + "_" + (l + 1) + ".txt");
+            OASInstancesWithTOU OASInstances1 = new OASInstancesWithTOU();
+            String instanceName = new String(".\\instances\\SingleMachineOASWithTOU\\" + orders[i] + "orders\\Tao" + Tao[j] + "\\R" + R[k]
+                    + "\\Dataslack_" + orders[i] + "orders_Tao" + Tao[j] + "R" + R[k] + "_" + (l + 1) + ".txt");
             OASInstances1.setData(instanceName, orders[i]);
             OASInstances1.getDataFromFile();
 
@@ -172,23 +170,23 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
               for (int t = 0; t < crossoverType.length; t++) {
                 for (int n = 0; n < mutationRate.length; n++) {
                   for (int o = 0; o < elitism.length; o++) {
-                    for (int p = 0; p < numberOfSalesmen.length; p++) {                        
+                    for (int p = 0; p < numberOfSalesmen.length; p++) {
+                      for (int ls = 0; ls < 2; ls++) {
+                        if (ls == 0) {
+                          applyLocalSearch = false;
+                        } else {
+                          applyLocalSearch = true;
+                        }
                         for (int q = 0; q < alpha.length; q++) {
                           for (int pop_Size = 0; pop_Size < populationsSize.length; pop_Size++) {
                             for (int r = 0; r < repeat; r++) {
-                              if (alpha[q] != 0) {
-                                applyLocalSearch = true;
-                              } else {
-                                applyLocalSearch = false;
-                              }                              
-                              
                               int _alpha = (int) Math.round(((double) orders[i] * alpha[q]));
-                              generations[0] = orders[i] * (numberOfSalesmen[p] - 1) * 2000 / populationsSize[pop_Size];
-                              singleMachineOAS_SGA TSP1 = new singleMachineOAS_SGA();
+                              generations[0] = orders[i] * (numberOfSalesmen[p] - 1) * 100 / populationsSize[pop_Size];
+                              singleMachineOAS_TPSGA_Co2 TSP1 = new singleMachineOAS_TPSGA_Co2();
                               TSP1.alpha = alpha[q];
                               TSP1.setParameter(crossoverRate[m], mutationRate[n], counter, elitism[o], generations[0],
-                                      crossoverType[t], numberOfSalesmen[p], OASInstances1.getSize(), instanceName,
-                                      OASInstances1.getR(), OASInstances1.getP(), OASInstances1.getD(), OASInstances1.getD_bar(), OASInstances1.getE(), OASInstances1.getW(), OASInstances1.getS(), populationsSize[pop_Size]);
+                                      crossoverType[t], numberOfSalesmen[p], OASInstances1.getSize(), instanceName, populationsSize[pop_Size],
+                                      OASInstances1.getP(), OASInstances1.getPower());
                               TSP1.setLocalSearchData(applyLocalSearch, _alpha);
                               TSP1.initiateVars();
                               TSP1.start();
@@ -197,6 +195,7 @@ public class singleMachineOAS_SGA extends mTSPSGATwoPart {
                             }
                           }
                         }
+                      }
                     }
                   }
                 }
